Time: 3 ms (68.00%), Space: 7.1 MB (18.55%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool judgeSquareSum(int c) {
-      int l=0;long long r=sqrt(c);
+      long long l=0;long long r=sqrt(c);
         while(l<=r){
             long long ans=l*l+r*r;
             if(ans==c)return true;
